+++ src/Garden.hpp
@@ -9,7 +9,7 @@ public:
     static std::vector<Garden *> gardens;
     static void add(Garden& g);
     
-    Garden(uint16_t x, uint16_t y, uint16_t width, uint16_t height, uint16_t tile_size);
+    Garden(float x, float y, uint16_t width, uint16_t height, uint16_t tile_size);
 
     void render();
     void surround_with_walls();
